requirements.txt
torch
torchvision
pillow
numpy

git+https://TOKEN_PLACEHOLDER@github.com/Pfeffabeissa/medilens-utils.git
